Add global option to pronounce words the moment they are translated
